Improve Similar Artists Diagnostic Output
This is looking for names that are similar. Basically it was looking for artists like "The Beatles" entered as "Beatles". Then I'd find and fix those.

HOWEVER -> Osees change their name and sometimes they are "Thee Oh Sees" and sometimes "Oh Sees". Also I have a album by "Girls" but have seen a band called "The Girls". Perhaps "Girls" are the same?

Regardless, both artists appear in the output OK, and these are the last that log for me now.