Add sort name property to Venue
"Thee Parkside" will still sort properly when the data is updated.

This allows "THEE" to be removed from the simplified name comparison for `LIBRARY_COMPARATOR`. "The" and "A" are still dropped when sorting names. "Oh Sees" & "Thee Oh Sees" were related, but not showing properly due to the simplified name sort.

Investigated by changing the settings.property `embedLinks` to false as well as passing `JAVA_OPTS='-Dweb.debug_output=true'` to the command line tool.